adding a content generator "system"
